perf(matmul): hash exit values, widen path coefficients, add matrix addition
Three changes, all driven by benchmarking the crate against the reference
C++ CFLOBDD on quantum circuits (GHZ, Bernstein-Vazirani, Deutsch-Jozsa,
QFT).

Hashed exit-value interning. Every product collapses the exits that
carry equal values, and that search was linear -- quadratic in the exit
count. Harmless for boolean or structured operands, which have a handful
of distinct values; ruinous for a Fourier-transformed state, which has
thousands. `perf` put 97.6% of a 16-qubit QFT in `substitute` and
`collapse`. `ValueSet` now scans while the list is short and switches to
hashing once it is not, keyed by the new `MatMulValue::dedup_key`:

    QFT, 16 qubits:  2341 ms -> 65 ms

The threshold matters -- hashing every lookup made the dense 256x256
matrix multiply slower, since it never has many distinct values. `f64`
keys on its bit pattern, normalised so that -0.0 and 0.0 agree with `==`
and NaN stays off the hashed path; a type with no key keeps the scan.

`i128` path coefficients. The deferred semiring counts how many products
coincide, and a Hadamard layer over n qubits makes that count 2^n, so
`i64` overflowed at 64 qubits -- exact arithmetic refusing to be wrong
rather than a blow-up. Widening doubles the reachable size and costs
~13% on the dense multiply. The reference implementation uses
`boost::multiprecision::cpp_int` here, which is why it reaches 256
qubits where this crate now stops at 64; a coefficient that promotes to
a bignum on overflow would close that gap.

`mk_matadd` and `mk_scale`. Anything that is not a product operator
needs addition: a controlled gate is `|0><0| (x) I + |1><1| (x) U`, two
Kronecker towers added. `mk_op_pair_map` cannot serve, since it demands
`Copy + Eq` and amplitudes are neither. Both route through the same
`collapse` as the products, so a sum that cancels or a scale by zero
lands on the constant diagram rather than merely evaluating to it.

`GcflobddT::count_nodes_and_edges` measures one diagram -- reachable
nodes and the connections leaving them -- as opposed to
`Context::node_count`, which counts everything interned. It is the
figure that compares to another implementation's node/edge counts.

+++ src/gcflobdd/matmul/map.rs
@@ -14,13 +14,18 @@ use std::cmp::Ordering;
 /// its cache) depends only on structure, and the same code serves every value
 /// type.
 ///
+/// Coefficients are `i128`: they count how many products coincide, which for a
+/// Hadamard-style operator is exponential in the qubit count (the reference C++
+/// implementation reaches for arbitrary-precision integers here). Overflow
+/// panics rather than wrapping.
+///
 /// Entries are kept sorted by `(i, j)` and any key whose coefficient reaches
 /// zero is dropped, so `PartialEq`/`Hash` are canonical and the empty map is
 /// the *only* spelling of zero. (The C++ implementation instead writes zero as
 /// a `(-1, -1)` sentinel that coexists with genuine zero coefficients, which is
 /// exactly the conflation the algorithm notes warn about.)
 #[derive(Clone, PartialEq, Eq, Hash, Debug, Default)]
-pub(crate) struct MatMulMap(Vec<(usize, usize, i64)>);
+pub(crate) struct MatMulMap(Vec<(usize, usize, i128)>);
 
 impl MatMulMap {
     /// The additive identity: an empty combination.
@@ -41,7 +46,7 @@ impl MatMulMap {
     }
 
     #[inline]
-    pub fn iter(&self) -> impl Iterator<Item = (usize, usize, i64)> + '_ {
+    pub fn iter(&self) -> impl Iterator<Item = (usize, usize, i128)> + '_ {
         self.0.iter().copied()
     }
 
@@ -78,7 +83,7 @@ impl MatMulMap {
     }
 
     /// Scale every coefficient by `coeff`.
-    pub fn scale(&self, coeff: i64) -> Self {
+    pub fn scale(&self, coeff: i128) -> Self {
         if coeff == 0 {
             return Self::zero();
         }
@@ -113,8 +118,8 @@ impl MatMulMap {
     }
 
     /// Sum adjacent entries that share a key, dropping the ones that cancel.
-    fn compacted(sorted: Vec<(usize, usize, i64)>) -> Self {
-        let mut out: Vec<(usize, usize, i64)> = Vec::with_capacity(sorted.len());
+    fn compacted(sorted: Vec<(usize, usize, i128)>) -> Self {
+        let mut out: Vec<(usize, usize, i128)> = Vec::with_capacity(sorted.len());
         for (i, j, c) in sorted {
             match out.last_mut() {
                 Some(last) if last.0 == i && last.1 == j => {
